Strip text resembling HTML tags from metadata
Each line of a metadata block has all HTML-like tags removed before
parsing.  Basically this removes all text contained between and
including '<' and '>', but it is not "greedy".  For example, the
metadata:

title : The <, ==, >, and <=> Operators

would be stripped to:

title : The , and  Operators

It would NOT be stripped to:

title : The  Operators

The metadata test have also been re-enabled.