Test for update working dir for existing page in repo with page file dir configured.
A wiki with a page file directory will not do a proper checkout during the update working directory due to the duplicate page file directory prepended to the path. The page file directory is duplicated at the beginning of the path because an existing page retrieved will already include it's full path, including the page file directory, in the page path.