Updating a page in a wiki with the page_file_dir configured will always leave a staged commit after the save.
A wiki with a page file directory configured will not do a proper checkout during the update working directory process, leaving a staged commit, due to a duplicate page file directory prepended to the path. An existing page retrieved during the update process will already include it's full path, including the page file directory, in the page path so it doesn't need to be added to the dir during the update.