Fix #154: Code block not correctly escaped when using pygments
For some reason code blocks used to default to simple <pre> tags
when the specified language was nil or bash. This caused html tags
in the code block to not get escaped since they were never passed
to pygments.

This fix defaults to the 'text' language if none is specified and
allows pygments to render bash code blocks since it supports it.