Don't expand filesystem paths
Paths starting with a `~` tilde character were getting expanded by the
call to `File.expand_path` in `BlobEntry.normalize_dir`. This can cause
an exception when the tilde is followed by an invalid username, which
makes the whole wiki unusable.

This overrides `BlobEntry.normalize_dir` so it doesn't expand tildes
anymore, and extends the tests a bit.

There's another call to `File.expand_path` in `File.get_disk_reference`
which could be exploited similarly, this one was just removed because it
doesn't seem to be needed.