Conditionally render "editable" heading classes
In the `gollum` gem, wikis are initialized with a setting
`allow_editing`. So far, this gem is unaware of that setting and does
nothing with it.

However, this gem does consider some headings in a document "editable"
and other headings not editable. It specifies a CSS class, "editable"
on any heading it considers editable.

This commit gives `gollum-lib` more context about whether a heading
should be editable or not based on the settings of the current `Wiki`.
The default value is `true`, but now we can set the `allow_editing`
setting to specify whether a heading in a document should be rendered
with the "editable" class or not.